+++ src/test/kotlin/no/nav/k9brukerdialogapi/ytelse/omsorgspengerutvidetrett/domene/OmsorgspengerUtvidetRettSøknadTest.kt
@@ -1,17 +1,20 @@
-package no.nav.k9brukerdialogapi.ytelse.omsorgspengerutvidetrett
+package no.nav.k9brukerdialogapi.ytelse.omsorgspengerutvidetrett.domene
 
 import no.nav.helse.dusseldorf.ktor.core.Throwblem
-import no.nav.k9brukerdialogapi.ytelse.omsorgspengerutvidetrett.domene.Barn
-import no.nav.k9brukerdialogapi.ytelse.omsorgspengerutvidetrett.domene.SøkerBarnRelasjon
-import no.nav.k9brukerdialogapi.ytelse.omsorgspengerutvidetrett.domene.Søknad
+import no.nav.k9brukerdialogapi.SøknadUtils
+import no.nav.k9brukerdialogapi.somJson
+import org.json.JSONObject
 import org.junit.jupiter.api.Assertions
+import org.skyscreamer.jsonassert.JSONAssert
+import java.time.ZoneId
+import java.time.ZonedDateTime
 import kotlin.test.Test
 import kotlin.test.assertTrue
 
-class OmsorgspengerUtvidetRettValideringTest {
+class OmsorgspengerUtvidetRettSøknadTest {
 
     @Test
-    fun `Skal ikke feile på gyldig søknad`() {
+    fun `Validering skal ikke feile på gyldig søknad`() {
         Søknad(
             språk = "nb",
             kroniskEllerFunksjonshemming = true,
@@ -29,7 +32,7 @@ class OmsorgspengerUtvidetRettValideringTest {
     }
 
     @Test
-    fun `Forvent feil dersom sammeAdresse er false og mangler samværsavtale`(){
+    fun `Forvent valideringsfeil dersom sammeAdresse er false og mangler samværsavtale`(){
         val feil = Assertions.assertThrows(Throwblem::class.java) {
             Søknad(
                 språk = "nb",
@@ -49,7 +52,7 @@ class OmsorgspengerUtvidetRettValideringTest {
     }
 
     @Test
-    fun `Forventer feil dersom harForståttRettigheterOgPlikter er false`(){
+    fun `Forventer valideringsfeil dersom harForståttRettigheterOgPlikter er false`(){
         val feil: String = Assertions.assertThrows(Throwblem::class.java){
             Søknad(
                 språk = "nb",
@@ -68,7 +71,7 @@ class OmsorgspengerUtvidetRettValideringTest {
     }
 
     @Test
-    fun `Forventer feil dersom harBekreftetOpplysninger er false`(){
+    fun `Forventer valideringsfeil dersom harBekreftetOpplysninger er false`(){
         val feil: String = Assertions.assertThrows(Throwblem::class.java){
             Søknad(
                 språk = "nb",
@@ -84,6 +87,52 @@ class OmsorgspengerUtvidetRettValideringTest {
             ).valider()
         }.message.toString()
         assertTrue(feil.contains("Opplysningene må bekreftes for å sende inn søknad."))
+    }
+
+    @Test
+    fun `Mapping av k9format blir som forventet`(){
+        val søknad = Søknad(
+            språk = "nb",
+            mottatt = ZonedDateTime.of(2020, 1, 2, 3, 4, 5, 6, ZoneId.of("UTC")),
+            barn = Barn(
+                norskIdentifikator = "02119970078",
+                fødselsdato = null,
+                aktørId = null,
+                navn = "Barn Barnsen"
+            ),
+            sammeAdresse = true,
+            relasjonTilBarnet = SøkerBarnRelasjon.FOSTERFORELDER,
+            kroniskEllerFunksjonshemming = true,
+            harForståttRettigheterOgPlikter = true,
+            harBekreftetOpplysninger = true
+        )
+        val faktiskK9Format = JSONObject(søknad.tilK9Format(SøknadUtils.søker).somJson())
+        val forventetK9Format = JSONObject(
+            """
+                {
+                  "språk": "nb",
+                  "mottattDato": "2020-01-02T03:04:05.000Z",
+                  "søknadId": "${søknad.søknadId}",
+                  "søker": {
+                    "norskIdentitetsnummer": "26104500284"
+                  },
+                  "ytelse": {
+                    "barn": {
+                      "fødselsdato": null,
+                      "norskIdentitetsnummer": "02119970078"
+                    },
+                    "kroniskEllerFunksjonshemming": true,
+                    "type": "OMP_UTV_KS"
+                  },
+                  "journalposter": [],
+                  "begrunnelseForInnsending": {
+                    "tekst": null
+                  },
+                  "versjon": "1.0.0"
+                }
+            """.trimIndent()
+        )
+        JSONAssert.assertEquals(forventetK9Format, faktiskK9Format, true)
     }
 
 }
